+++ src/CommandBus/CanNotDispatchCommandException.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright   Copyright (c) 2020 - 2023 Communitales GmbH (http://www.communitales.com/)
+ * @copyright   Copyright (c) 2020 - 2024 Communitales GmbH (http://www.communitales.com/)
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -9,6 +9,7 @@
 namespace Communitales\Component\CommandBus;
 
 use RuntimeException;
+
 use function sprintf;
 
 class CanNotDispatchCommandException extends RuntimeException