+++ src/CommandBus/Handler/Result/AbstractResult.php
@@ -1,6 +1,7 @@
 <?php
+
 /**
- * @copyright   Copyright (c) 2020 - 2023 Communitales GmbH (https://www.communitales.com/)
+ * @copyright   Copyright (c) 2020 - 2024 Communitales GmbH (https://www.communitales.com/)
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,15 +13,7 @@ use Communitales\Component\StatusBus\StatusMessage;
 
 abstract class AbstractResult implements CommandHandlerResultInterface
 {
-    private ?StatusMessage $statusMessage;
-
-    public function __construct(?StatusMessage $statusMessage = null)
+    public function __construct(public readonly ?StatusMessage $statusMessage = null)
     {
-        $this->statusMessage = $statusMessage;
-    }
-
-    public function getStatusMessage(): ?StatusMessage
-    {
-        return $this->statusMessage;
     }
 }