+++ src/CommandBus/Handler/Result/CommandHandlerResultInterface.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright   Copyright (c) 2020 - 2023 Communitales GmbH (https://www.communitales.com/)
+ * @copyright   Copyright (c) 2020 - 2024 Communitales GmbH (https://www.communitales.com/)
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -15,5 +15,4 @@ use Communitales\Component\StatusBus\StatusMessage;
  */
 interface CommandHandlerResultInterface
 {
-    public function getStatusMessage(): ?StatusMessage;
 }