+++ src/CommandBus/Handler/Result/CommandResultException.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * @copyright   Copyright (c) 2020 - 2023 Communitales GmbH (https://www.communitales.com/)
+ * @copyright   Copyright (c) 2020 - 2024 Communitales GmbH (https://www.communitales.com/)
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
@@ -12,25 +12,17 @@ use RuntimeException;
 use Throwable;
 
 /**
- * An exception representing a result. In some cases this may be useful.
+ * An exception representing an CommandHandlerResult.
+ * This is useful if you want to return a result from a sub method of a handler method.
  */
 class CommandResultException extends RuntimeException
 {
-    private CommandHandlerResultInterface $commandResult;
-
     public function __construct(
-        CommandHandlerResultInterface $commandResult,
+        public readonly CommandHandlerResultInterface $commandResult,
         string $message = '',
         int $code = 0,
         Throwable $previous = null
     ) {
-        $this->commandResult = $commandResult;
         parent::__construct($message, $code, $previous);
     }
-
-    public function getCommandResult(): CommandHandlerResultInterface
-    {
-        return $this->commandResult;
-    }
-
 }